rename examples to make it more understanding
