+++ examples/README.md
@@ -13,8 +13,8 @@ cargo run --package example-{name of example}
 - [axum](./axum) - Use autometrics to instrument HTTP handlers using the `axum` framework
 - [custom-metrics](./custom-metrics/) - Define your own custom metrics alongside the ones generated by autometrics (using any of the metrics collection crates)
 - [exemplars-tracing](./exemplars-tracing/) - Use fields from `tracing::Span`s as Prometheus exemplars
-- [opentelemetry-push](./opentelemetry-push/) - Push metrics to an OpenTelemetry Collector via the OTLP gRPC protocol
-- [opentelemetry-push-http](./opentelemetry-push-http/) - Push metrics to an OpenTelemetry Collector via the OTLP HTTP protocol using the Autometrics provided interface
+- [opentelemetry-push](./opentelemetry-push/) - Push metrics to an OpenTelemetry Collector via the OTLP HTTP or gRPC protocol using the Autometrics provided interface
+- [opentelemetry-push-custom](./opentelemetry-push-custom/) - Push metrics to an OpenTelemetry Collector via the OTLP gRPC protocol using custom options
 
 ## Full Example
