+++ examples/opentelemetry-push-custom/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "example-opentelemetry-push-custom"
+version = "0.0.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+autometrics = { path = "../../autometrics", features = ["opentelemetry-0_20"] }
+autometrics-example-util = { path = "../util" }
+# Note that the version of the opentelemetry crate MUST match
+# the version used by autometrics
+opentelemetry = { version = "0.20", features = ["metrics", "rt-tokio"] }
+opentelemetry-otlp = { version = "0.13", features = ["tonic", "metrics"] }
+opentelemetry-semantic-conventions = { version = "0.12.0" }
+tokio = { version = "1", features = ["full"] }
