+++ examples/opentelemetry-push-custom/otel-collector-config.yml
@@ -1,8 +1,6 @@
 receivers:
   otlp:
     protocols:
-      http:
-        endpoint: 0.0.0.0:4318
       grpc:
         endpoint: 0.0.0.0:4317
