--- examples/opentelemetry-push-http/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "example-opentelemetry-push-http"
-version = "0.0.0"
-publish = false
-edition = "2021"
-
-[dependencies]
-autometrics = { path = "../../autometrics", features = ["opentelemetry-0_20", "otel-push-exporter-http", "otel-push-exporter-tokio"] }
-autometrics-example-util = { path = "../util" }
-tokio = { version = "1", features = ["full"] }
