+++ examples/opentelemetry-push/otel-collector-config.yml
@@ -1,6 +1,8 @@
 receivers:
   otlp:
     protocols:
+      http:
+        endpoint: 0.0.0.0:4318
       grpc:
         endpoint: 0.0.0.0:4317
